+++ services/offerprice.js
@@ -1,6 +1,6 @@
 function findOfferPrice(price){
     const offer = 50 ;
-    const offerprice = (parseFloat(price) * offer) / 100;
+    const offerprice = (parseInt(price) * offer) / 100;
     return offerprice;
 }
 module.exports = findOfferPrice;
